✨ Feat: 그룹별 pdf 조회 API 구현
