+++ src/main/java/com/likelion/server/domain/pdf/repository/PdfRepository.java
@@ -3,4 +3,8 @@ package com.likelion.server.domain.pdf.repository;
 import com.likelion.server.domain.pdf.entity.Pdf;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PdfRepository extends JpaRepository<Pdf, Long> {}
+import java.util.List;
+
+public interface PdfRepository extends JpaRepository<Pdf, Long> {
+    List<Pdf> findAllByGroupId(Long groupId);
+}
